requirements.txt
torch==2.2.0
gymnasium==0.29.1
stable-baselines3==2.3.0
pandas==2.0.3
numpy==1.24.3
pyyaml==6.0.1
plotly==5.18.0
matplotlib==3.8.2
ta==0.11.0
pytest==7.4.3
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
python-dateutil==2.8.2
typing-extensions==4.9.0 